refactor: block order update hook
