+++ apps/journeys-admin/src/components/Editor/Slider/Content/Canvas/QuickControls/MoveBlock/MoveBlock.tsx
@@ -1,4 +1,3 @@
-import { gql, useMutation } from '@apollo/client'
 import IconButton from '@mui/material/IconButton'
 import { ReactElement } from 'react'
 
@@ -8,23 +7,10 @@ import { searchBlocks } from '@core/journeys/ui/searchBlocks'
 import ChevronDownIcon from '@core/shared/ui/icons/ChevronDown'
 import ChevronUpIcon from '@core/shared/ui/icons/ChevronUp'
 
-import { BlockOrderUpdate } from '../../../../../../../../__generated__/BlockOrderUpdate'
-
-export const BLOCK_ORDER_UPDATE = gql`
-  mutation BlockOrderUpdate($id: ID!, $journeyId: ID!, $parentOrder: Int!) {
-    blockOrderUpdate(
-      id: $id
-      journeyId: $journeyId
-      parentOrder: $parentOrder
-    ) {
-      id
-      parentOrder
-    }
-  }
-`
+import { useBlockOrderUpdateMutation } from '../../../../../../../libs/useBlockOrderUpdateMutation'
 
 export function MoveBlock(): ReactElement {
-  const [blockOrderUpdate] = useMutation<BlockOrderUpdate>(BLOCK_ORDER_UPDATE)
+  const [blockOrderUpdate] = useBlockOrderUpdateMutation()
   const { journey } = useJourney()
   const {
     state: { selectedBlock, selectedStep }
